widget: state と純関数クラスタを 4 モジュールに分割する (#109 R-3)
widget/src/index.js (1729 行) から挙動を変えずにコードを一字一句そのまま
移動して 4 モジュールを新設する (verbatim move):

- state.js: DEFAULTS と state (オブジェクト参照共有。state への再代入は
  無く全参照がプロパティ読み書きのため挙動不変)
- screenshot.js: captureScreenshot / buildScreenshotSvg /
  screenshotOverlayFor / collectReadableStyles / visibleBackground /
  snapshotClassAttr / serializeAsXhtml / renderScreenshotOverlay /
  byteLength / base64Encode
- persistence.js: FEEDBACK_STORAGE_VERSION / loadStoredReviewer /
  saveReviewer / loadPersistedFeedback / persistFeedbackList /
  clearPersistedFeedback / feedbackStorageEnvelope /
  serializeFeedbackForStorage / normalizePersistedFeedback /
  isMatchingFeedbackEnvelope
- payload.js: PAYLOAD_SCHEMA_VERSION / buildPayload / generateFeedbackId
  (呼び出し元が buildPayload のみであることを grep で確認)

restorePersistedFeedback は index.js 残留関数 (removeCommittedMarkers /
restoreFeedbackMarkers) を呼ぶ orchestrator のため index.js に残す。
Slack 直送系 (postSlackWebhook ほか) は #97 で削除予定のため移動しない。
safeFilePart は R-1 スライスと競合させないため触らない。

新モジュールは全て「葉」(index.js への逆依存なし) で、dist の差分は
モジュール境界コメント・IIFE ラッパ・並び順・import 行のみ。全 1912 行の
関数本体テキストがバイト一致することを機械照合で確認済み。

node 環境で動く純関数に単体テストを追加:
- test/widget-persistence.test.js: serialize/normalize の round-trip と
  envelope 一致判定
- test/widget-screenshot.test.js: byteLength / base64Encode
